+++ glym-app/src/features/auth/index.js
@@ -1,1 +1,3 @@
 export { default as AuthForm } from "./components/AuthForm";
+export { default as SignUpForm } from "./components/SignUpForm";
+export { default as SignUpInput } from "./components/SignUpInput";
